+++ android/src/main/kotlin/com/raghavg1999/flutter_ffmpeg_utils/FfmpegUtilsPlugin.kt
@@ -10,6 +10,7 @@ import com.arthenica.ffmpegkit.FFmpegKit
 import com.arthenica.ffmpegkit.FFmpegSession
 import com.arthenica.ffmpegkit.ReturnCode
 
+
 /** FfmpegUtilsPlugin */
 class FfmpegUtilsPlugin: FlutterPlugin, MethodCallHandler {
   /// The MethodChannel that will the communication between Flutter and native Android
@@ -25,43 +26,12 @@ class FfmpegUtilsPlugin: FlutterPlugin, MethodCallHandler {
 
   override fun onMethodCall(call: MethodCall, result: Result) {
     if (call.method == "executeFFmpeg") {
-      val arguments = call.arguments as List<String>
-            val exitCode =     executeFFmpeg(arguments)
-            if (exitCode == 0) {
-                result.success("FFmpeg command executed successfully!")
-            } else {
-                result.error("FFMPEG_ERROR", "FFmpeg command failed with exit code $exitCode", null)
-            }
+        val command = call.arguments as List<String>
+        FfmpegHelper.executeFFmpeg(command, result)
     } else {
       result.notImplemented()
     }
   }
-
-   private fun executeFFmpeg(commandList: List<String>): Int {
-        // Join the command list into a single string
-        val command = commandList.joinToString(" ")
-
-        // Execute the FFmpeg command (blocking call)
-        val session: FFmpegSession = FFmpegKit.execute(command)
-
-        // Check the return code to see if the command was successful
-        when {
-            ReturnCode.isSuccess(session.returnCode) -> {
-                println("✅ FFmpeg command executed successfully.")
-                return 0;
-            }
-            ReturnCode.isCancel(session.returnCode) -> {
-                println("⚠️ FFmpeg command was cancelled.")
-                return -1;
-            }
-            else -> {
-                println("❌ FFmpeg command failed with return code: ${session.returnCode}")
-                println("Output: ${session.output}")
-                println("Error: ${session.failStackTrace}")
-                return -1;
-            }
-        }
-    }
 
   override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)